# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RAMP UP — Engineering Onboarding Portal API
  description: |
    Tài liệu đặc tả API (OpenAPI Specification) cho hệ thống **RAMP UP** (Glinteco e-Learning).
    Hệ thống cung cấp cổng thông tin đào tạo nội bộ gamified dành cho các kỹ sư mới gia nhập Glinteco.
    
    ### Quy ước chung:
    * **Base URL:** `/api/v1` (mặc định môi trường development: `https://api.glinteco-elearning.dev/api/v1`)
    * **Xác thực:** Ngoại trừ các route Auth public, tất cả các request yêu cầu header `Authorization: Bearer <accessToken>`.
    * **Định dạng thời gian:** Chuẩn ISO-8601 UTC (ví dụ: `2026-06-16T15:16:56Z`).
    * **Phân trang danh sách (Pagination):** Sử dụng Keyset/Cursor Pagination với định dạng phản hồi chung:
      ```json
      {
        "data": [...],
        "nextCursor": "<opaque_string_or_null>",
        "hasMore": true
      }
      ```
    * **Quản lý lỗi (Error Envelopes):** Các lỗi trả về định dạng chuẩn:
      ```json
      {
        "statusCode": 400,
        "message": ["email must be a valid email"],
        "error": "Bad Request"
      }
      ```
  version: 1.0.0
servers:
  - url: https://api.glinteco-elearning.dev/api/v1
    description: Môi trường Development
paths:
...
  /auth/register:
    post:
      summary: Đăng ký tài khoản mới (Public)
      description: Tạo tài khoản learner mới với email, password và họ tên.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: Ada Lovelace
                email:
                  type: string
                  format: email
                  example: ada@acme.dev
                password:
                  type: string
                  format: password
                  example: hunter2-very-strong
      responses:
        '201':
          description: Đăng ký thành công và tự động tạo phiên đăng nhập.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
                    example: eyJhbGciOiJIUzI1Ni…
                  refreshToken:
                    type: string
                    example: rt_a1b2c3d4e5f6
                  expiresIn:
                    type: integer
                    description: Thời hạn token tính bằng giây (ví dụ 900 = 15 phút).
                    example: 900
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email đã được sử dụng.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: Đăng nhập bằng Email & Password (Public)
      description: Xác thực người dùng bằng email và mật khẩu để lấy access token và refresh token.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: mina@acme.dev
                password:
                  type: string
                  format: password
                  example: password123
                rememberMe:
                  type: boolean
                  default: false
                  example: true
      responses:
        '200':
          description: Đăng nhập thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
                    example: eyJhbGciOiJIUzI1Ni…
                  refreshToken:
                    type: string
                    example: rt_9z8y7x6w
                  expiresIn:
                    type: integer
                    example: 900
        '401':
          description: Sai thông tin đăng nhập.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/google:
    post:
      summary: Đăng nhập / Đăng ký qua Google OAuth (Public)
      description: Sử dụng ID Token gửi từ client sau khi xác thực qua Google. Tài khoản mới sẽ được tự động kích hoạt vai trò `learner`.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - idToken
              properties:
                idToken:
                  type: string
                  example: google-oauth2-id-token-here
      responses:
        '200':
          description: Xác thực Google thành công. Trả về thông tin phiên giống như `/auth/login`.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
                    example: eyJhbGciOiJIUzI1Ni…
                  refreshToken:
                    type: string
                    example: rt_9z8y7x6w
                  expiresIn:
                    type: integer
                    example: 900
        '400':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      summary: Làm mới Access Token (Public)
      description: Gửi refresh token hiện tại để nhận access token mới cùng refresh token mới.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  example: rt_9z8y7x6w
      responses:
        '200':
          description: Cấp token mới thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: eyJhbGciOiJIUzI1Ni…
                  refreshToken:
                    type: string
                    example: rt_new1234
                  expiresIn:
                    type: integer
                    example: 900
        '401':
          description: Refresh token không hợp lệ hoặc đã hết hạn.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      summary: Đăng xuất (Authenticated)
      description: Thu hồi refresh token hiện tại để hủy phiên làm việc.
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Đăng xuất thành công. Không có dữ liệu trả về.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/forgot-password:
    post:
      summary: Yêu cầu đặt lại mật khẩu (Public)
      description: Gửi email liên kết đặt lại mật khẩu nếu tài khoản tồn tại trong hệ thống.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: mina@acme.dev
      responses:
        '202':
          description: Yêu cầu đã được tiếp nhận.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: If the account exists, a reset link has been sent.
  /auth/reset-password:
    post:
      summary: Đặt lại mật khẩu mới (Public)
      description: Sử dụng token nhận được từ email để thiết lập mật khẩu mới.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                  example: reset_tok_abc123
                password:
                  type: string
                  format: password
                  example: new-strong-password
      responses:
        '200':
          description: Đổi mật khẩu thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password updated. Please sign in.
        '400':
          $ref: '#/components/responses/ValidationError'
  /auth/me:
    get:
      summary: Lấy thông tin tài khoản hiện tại (Authenticated)
      description: Lấy thông tin chi tiết của người dùng đang đăng nhập dựa trên JWT access token.
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # =========================================================================
  # 2. USERS & PROFILE
  # =========================================================================
  /users:
    get:
      summary: Danh sách người dùng (Admin)
      description: Lấy danh sách người dùng trong hệ thống hỗ trợ tìm kiếm và lọc. Chỉ dành cho vai trò `admin`.
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - name: cohortId
          in: query
          description: Lọc danh sách theo Cohort ID
          required: false
          schema:
            type: string
            example: c_spring2026
        - name: role
          in: query
          description: Lọc theo vai trò (learner hoặc admin)
          required: false
          schema:
            type: string
            enum: [learner, admin]
            example: learner
        - name: q
          in: query
          description: Tìm kiếm theo họ tên hoặc email
          required: false
          schema:
            type: string
            example: Mina
        - name: limit
          in: query
          description: Giới hạn số lượng bản ghi trả về (tối đa 50)
          required: false
          schema:
            type: integer
            maximum: 50
            default: 20
            example: 20
        - name: cursor
          in: query
          description: Cursor phục vụ phân trang danh sách
          required: false
          schema:
            type: string
            example: eyJjcmVhdGVkQXQiOiIyMDI2…
      responses:
        '200':
          description: Thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserSummary'
                  nextCursor:
                    type: string
                    nullable: true
                    example: eyJjcmVhdGVkQXQiOiIyMDI2…
                  hasMore:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{id}:
    get:
      summary: Lấy chi tiết thông tin một người dùng (Authenticated)
      description: |
        Lấy thông tin profile chi tiết của một kỹ sư.
        * Quy tắc phân quyền:* `learner` chỉ được phép lấy chính profile của họ, `admin` được phép lấy bất kỳ ai.
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID của người dùng cần xem
          required: true
          schema:
            type: string
            example: u_mina01
      responses:
        '200':
          description: Thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Không đủ quyền truy cập (Learner xem profile người khác).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Không tìm thấy người dùng.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me:
    patch:
      summary: Cập nhật thông tin profile cá nhân (Authenticated)
      description: Cho phép người dùng tự chỉnh sửa một số thuộc tính profile cá nhân như chức danh (title) hoặc màu đại diện (avatarHue).
      tags:
        - Users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Senior Frontend Engineer
                avatarHue:
                  type: integer
                  minimum: 0
                  maximum: 360
                  example: 200
      responses:
        '200':
          description: Cập nhật thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: u_mina01
                  title:
                    type: string
                    example: Senior Frontend Engineer
                  avatarHue:
                    type: integer
                    example: 200
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users/me/stats:
    get:
      summary: Thống kê tiến độ & Gamification cá nhân (Learner)
      description: Lấy bộ dữ liệu thống kê tổng hợp dùng để hiển thị trên màn hình dashboard của học viên (Level, XP, Streak, Tỷ lệ hoàn thành, số bài tập, bookmark).
      tags:
        - Users
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDashboardStats'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Gửi access token dạng 'Bearer <Token>' trong header Authorization.
  responses:
    ValidationError:
      description: Dữ liệu đầu vào không hợp lệ (lỗi validate DTO).
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 400
              message:
                type: array
                items:
                  type: string
                example: ["email must be a valid email", "password must be longer than or equal to 6 characters"]
              error:
                type: string
                example: Bad Request
    UnauthorizedError:
      description: Chưa đăng nhập hoặc Token hết hạn/không hợp lệ.
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 401
              message:
                type: string
                example: Unauthorized
              error:
                type: string
                example: Unauthorized
    ForbiddenError:
      description: Người dùng không có quyền truy cập chức năng này (sai role).
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 403
              message:
                type: string
                example: Forbidden resource
              error:
                type: string
                example: Forbidden
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Yêu cầu không hợp lệ
        error:
          type: string
          example: Bad Request
    User:
      type: object
      properties:
        id:
          type: string
          example: u_8f3a1c20
        name:
          type: string
          example: Ada Lovelace
        email:
          type: string
          example: ada@acme.dev
        role:
          type: string
          enum: [learner, admin]
          example: learner
        level:
          type: integer
          example: 1
        xp:
          type: integer
          example: 0
        cohortId:
          type: string
          nullable: true
          example: c_spring2026
    UserSummary:
      type: object
      properties:
        id:
          type: string
          example: u_mina01
        name:
          type: string
          example: Mina Okonkwo
        role:
          type: string
          example: learner
        level:
          type: integer
          example: 3
        xp:
          type: integer
          example: 1240
        completion:
          type: integer
          description: Tỷ lệ hoàn thành toàn bộ lộ trình (%)
          example: 46
        avatarHue:
          type: integer
          description: Mã màu HSL Hue đại diện đại diện
          example: 162
    UserDetail:
      type: object
      properties:
        id:
          type: string
          example: u_mina01
        name:
          type: string
          example: Mina Okonkwo
        email:
          type: string
          example: mina@acme.dev
        role:
          type: string
          example: learner
        title:
          type: string
          example: Frontend Engineer
        avatarHue:
          type: integer
          example: 162
        level:
          type: integer
          example: 3
        xp:
          type: integer
          example: 1240
        streakDays:
          type: integer
          example: 6
        cohortId:
          type: string
          example: c_spring2026
        joinedAt:
          type: string
          format: date-time
          example: 2026-04-02T09:00:00Z
    UserDashboardStats:
      type: object
      properties:
        level:
          type: integer
          example: 3
        xp:
          type: integer
          example: 1240
        xpThisWeek:
          type: integer
          example: 180
        streakDays:
          type: integer
          example: 6
        overallCompletion:
          type: integer
          example: 46
        tracks:
          type: object
          properties:
            completed:
              type: integer
              example: 2
            total:
              type: integer
              example: 5
        exercises:
          type: object
          properties:
            approved:
              type: integer
              example: 1
            total:
              type: integer
              example: 3
            awaitingReview:
              type: integer
              example: 1
        savedDocs:
          type: object
          properties:
            total:
              type: integer
              example: 12
            unread:
              type: integer
              example: 3
...